ZDC_SiPMonTile.xml: name rotation angle after Y-axis, not Z (#545)
Corrected which axis the ZDC_SiPMonTile is rotated about. This axis was
set to z accidentally, but this pull request changes this to the y axis.

### Briefly, what does this PR introduce?
Makes the rotation of the ZDC SiPM-on-tile be along the y axis, rather
than the z axis

### What kind of change does this PR introduce?
- [X] Bug fix (issue #544)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No breaking changes. This will modify the output of any code that uses
the output of ZDC SiPM-on-tile simulations
### Does this PR change default behavior?
Since the proposed SiPM-on-tile ZDC is not yet a part of the default
configuration, this will not change the default behavior