feat - Increase number of W/Sc layers in forward HCal insert from 4 to 10. (#547)
### Briefly, what does this PR introduce?
There are currently 4 layers of W/Sc in the forward HCal insert to match
the layering of the surrounding LFHCAL. The correct layer structure for
the HCal insert should be 1 front layer of Steel/Sc, 10 layers of W/Sc,
54 layers of Steel/Sc, 1 back layer of steel.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X] New feature (issue #546 )
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [X] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
None.
### Does this PR change default behavior?
Only the changes to the HCal insert listed above.