Update Outer MPGD barrel to match detector parameter table (#594)
### Briefly, what does this PR introduce?
Updates the outer MPGD barrel layer (the one integrated with the DIRC)
to match the detector parameter table

### What kind of change does this PR introduce?
- [x ] Bug fix (issue #568 )
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
This will probably cause an overlap with the DIRC volume, since the
updated DIRC volume is not accounted for in this branch.
### Does this PR change default behavior?
Will likely trigger overlap errors between the DIRC and MPGD layer