reduced vessel rmax2 to 180 cm (#592)
### Briefly, what does this PR introduce?
The updated geometry table requires the vessel rmax2 to be at 180 cm,
but the designed communicated by Marco Contalbrigo requires the sensor
box extended up to 185 cm. Also it requires the dRICH vessel length to
be 122 cm (currently not implemented) .
Marco is in communication with Tanja et al. to make sure the parameters
are fine and afterwards this draft PR will be changed to the final PR.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [X ] Changes have been communicated to collaborators
     Marco Contalbrigo and @c-dilks are communicated 
### Does this PR introduce breaking changes? What changes might users
need to make to their code?
It should not make any difference. But if the barrel calorimeters are
pushed downwards the overlap tests may fail.
### Does this PR change default behavior?
It should not make any difference. But if the barrel calorimeters are
pushed downwards the overlap tests may fail. Other than this the dRICH
performance should not change.

---------

Co-authored-by: Chao Peng <pengc2010@gmail.com>