ZDC LYSO + SiPM-on-tile (#610)
### Briefly, what does this PR introduce?
Adds ZDC_SiPMonTile and ZDC_Crystal_LYSO to the includes of the
compact/far_forward.xml, replacing the include for the deprecated
ZDC.xml file.
Also adds the file ZDC_Crystal_LYSO.xml, and a configuration containing
only the SiPM-on-tile and LYSO parts of the ZDC (and nothing else)

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X] New feature (issue #609 )
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
- [X] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
yes. If users want to use the old version of the ZDC (for instance to
compare behavior of the old version vs. the new one LYSO+SiPM-on-tile),
they must include it in their configuration, instead of including it
through the far_forward.xml
### Does this PR change default behavior?
yes.

---------

Co-authored-by: Barak Schmookler <bschmookler1@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>