Update ZDC_Crystal_LYSO.xml:  Use readout name EcalFarForwardZDCHits (#612)
Changed name of hits to EcalFarForwardZDCHits, as requested by @veprbl
when reviewing pull request 1221 in EICrecon.

### Briefly, what does this PR introduce?
Changes the name of the output bank for the LYSO section of the ZDC in
order to match the convention used for the SiPM-on-tile section

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [X] Other:  Changing the name of output banks to match the 
Change of the name of the readout output of the ZDC LYSO

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
- [X] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
Users will need to change their code to use the new name of the ZDC
Crystal LYSO detector, which is EcalFarForwardZDCHits
### Does this PR change default behavior?
yes

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>