Adding craterlake w/o barrel hcal to facilitate overlap checks (#633)
### Briefly, what does this PR introduce?

This adds an ePIC craterlake configuration w/o the barrel HCAL in order
to facilitate splitting the overlap checks. The barrel HCAL takes a very
long time for overlap checks due to the tessellated geometry. so
splitting it out should allow the overlpa checks to complete within the
6 hour time limit imposed by GitHub.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

None 

### Does this PR change default behavior?

No, it adds a new configuration.