feat: add the inter-MPGD/DIRC aluminum support rails (#521)
### Briefly, what does this PR introduce?
This adds the inter-MPGD/DIRC aluminum support rails. These are solid
aluminum (for now), and somewhat arbitrarily placed due to lack of info
on their position relative to the DIRC or MPGD. They are somewhat
arbitrarily added to the DIRC as opposed to the MPGD, but I assume
ultimately both will need to be connected with another bracket (of
minimal X0).

The aluminum rails are approximated as a trapezoid that ignores the top
360/48 angle, and that extends to the bottom of the round rail (which
isn't implemented here).


![image](https://github.com/eic/epic/assets/4656391/4f50a2c6-f246-46d5-8cbe-f5e2757da942)

![image](https://github.com/eic/epic/assets/4656391/8446a327-91cc-4149-a5e0-ae537c5da5b6)

![image](https://github.com/eic/epic/assets/4656391/a8ab5bb9-9d3b-4bec-8e31-83cd1cf609e2)

TODO:
- [ ] fix the inevitable overlaps...

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators @mariakzurek
@sly2j

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.