Updated HCAL parameters and template for comparison with detector table (#588)
### Briefly, what does this PR introduce?


### What kind of change does this PR introduce?
- [X] Bug fix (issue #562 )
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
None. The changes here are contained to the barrel HCAL geometry alone,
and then only the size of the outer HCAL mother volume. The only actual
effect is on the outer HCAL mother volume, which was reduced in size.

### Does this PR change default behavior?
No. The actual outer HCAL detector geometry is unchanged. 

NOTE: This review found that some of the parameters in the Detector
Table are incorrect. These changes will allow the proper values to be
calculated from the simulation input, but they won't match the Detector
Table until that table is updated.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Chao Peng <pengc2010@gmail.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>