feat: export simulation geometry to CAD models (STEP) (#667)
### Briefly, what does this PR introduce?
This PR exports the geometry to STEP files for CAD import and
dimensioning. There are no real settings to tweak, but to increase the
level at which volumes are exported, hints can be added to the yaml
file. Detectors can also be removed (currently only the ACTS
subassemblies are removed since they don't do anything).

There are 'issues' that need to be addressed, but they have to be
addressed in npsim:
- units are off (npdet_to_step use the cm value from TGeo and writes
them as inch values in STEP) https://github.com/eic/npsim/issues/20,
- some shapes cannot be exported; not sure yet which ones and it makes
more sense to bypass them in npsim
https://github.com/eic/npsim/issues/21.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>