Add new ff vacuum structure (#665)
### Briefly, what does this PR introduce?

This introduces a completely re-vamped far-forward beam pipe and vacuum
system, which includes a neutral particle exit window, boxes to house
the Roman pots, etc. and fills the volumes with vacuum.

There are a few additional updates which will be needed once some other
issues are resolved, but the overall layout will not drastically change.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [x] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

I have run the overlap checker (no overlaps), and done a few additional
visual checks + acceptance checks.

There are still some comments and missing "dd4hep::" constants in the
code - please ignore these for now. The comments are for my future
updates, and the missing constants are intentional, as all of the
drawings are in "mm", and i want to keep the exact numbers as they are
to make it easier to make updates.

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

Kind of - it drastically changes the behavior of particles which go to
the ZDC (a proper exit window is now there), and it will alter the
reconstruction of forward protons due to the shift of the RP planes (a
PR in EICrecon will follow with updated matrices).

It should be noted that the exit window really only covers up to about
3.8 mrad in neutral particle acceptance. This is due to the physical
constraints of fitting the window alongside the proton beam line and not
impacting the proton acceptance.

### Does this PR change default behavior?

Yes, especially for neutrons and photons in the ZDC.

---------

Co-authored-by: Alexander Jentsch <ajentsch@bnl.gov>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>