Extended configs tested by check-tracking-geometry (#697)
### Briefly, what does this PR introduce?
Adds tests to check if more configurations are pass the
check-tracking-geometry check.

Initially this is before the epic_ip6 has been updated so should fail
the test but probably initially won't. This should be followed by a
change to acts_geo_check to cover the case or a request to change the
logger level of messages which should cause it to fail.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
No