fix: export all configs to STEP, use `part --level 1 HcalBarrel` (#702)
### Briefly, what does this PR introduce?
Due to limitations in npdet_to_step, we disabled the HcalBarrel when
exporting to STEP. With npsim 1.4.0, the issue with the HcalBarrel was
[resolved](https://github.com/eic/npsim/pull/23).

This PR re-enables the HcalBarrel in npdet_to_step conversion, and it
simplifies the logic to export all detector configs (for consistency and
predictability).

TODO:
- [x] restart CI pipeline when eic-shell nightly rolls out with
npsim-1.4.0

### What kind of change does this PR introduce?
- [x] Bug fix (issue: HcalBarrel disabled in convert-to-step)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.