fix: update tracker support and services cones per project CAD info (#661)
### Briefly, what does this PR introduce?
This PR adds the outer support/services cone and places the cones and
barrels in their actual intended positions. I'm marking this as a bugfix
since the placement was completely wrong (with tracks in some regions
going through two support cylinders...).

TODO:
- [x] #663
- [x] #673
- this is not really a dependency, but if the radius of that system
increases, we want to figure it out before putting in the services
- [x] #668 
- [x] #683
- [x] #684

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators @sly2j 

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
Yes, it adds the project's estimate of the equivalent aluminum thickness
of the services.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>