Modified Barrel TOF geometry to add dead spaces between sensors and between support ring structure (#787)
### Briefly, what does this PR introduce?
The BTOF active volume is modified to be more realistic. Now there are
dead spaces between each sensors on a stave and extra dead space is
reserved in the center of the BTOF to make space for the central
supporting ring structure.
Variables relevant to the dead spaces, sensor sizes and the width of the
ring are adjustable from tof_barrel.xml.

![image](https://github.com/user-attachments/assets/7bd98ce5-5f2d-4ae9-9fcb-8b878973e9b4)


### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [x] Other: BTOF geometry is modified to accommodate individual
sensors.

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

### Does this PR change default behavior?

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>