Add functionality to track in a BeamPipeChain (#768)
### Briefly, what does this PR introduce?
An option had been added to include sensitive elements at the start
and/or end of beam pipe vacuum. This allows easy tracking of the shape
and position of beam electrons/protons to compare with accelerator group
and stand alone simulations. The segment in which scattered electrons of
different kinematics can also be studied/verified.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
No

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>