Update sensor geometry for rp (#791)
### Briefly, what does this PR introduce?

This PR introduces changing RP geometry for the various beam energies,
and updates the sensor size for the AC-LGADs to what we are planning to
build.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ x] Tests for the changes have been added (see PDF)
- [ x] Documentation has been added / updated
- [x ] Changes have been communicated to collaborators

Changes will be presented at meetings on Monday and Tuesday (Oct 14th
and 15th)

See slides here:
https://www.dropbox.com/scl/fi/43mhwxeyvluku0w1ojqfr/new_roman_pots_geometry_10_9_2024.pdf?rlkey=w0wc0gzrk4d03e8uh9w8s0tq5&dl=0

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No changes to user code needed, but users need to be careful to select
the correct geometry when running dd4hep AND EICrecon.
(https://chat.epic-eic.org/main/pl/rptqecge1jyctkkwa1454zy1mw)

I will hopefully address this issue in a PR immediately following this
one.

### Does this PR change default behavior?

Yes, the low-pT acceptances at the Roman pots will be much closer to
reality now, and will vary with beam energy, as they should.

---------

Co-authored-by: Alexander Jentsch <ajentsch@bnl.gov>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>