fix(ci): don't trigger geoviewer directly (#794)
### Briefly, what does this PR introduce?
Now that we trigger the geoviewer from the container job, along with
other benchmarks, we don't need to trigger it directly (and it doesn't
work to trigger it directly). E.g.
https://eicweb.phy.anl.gov/EIC/benchmarks/geoviewer/-/jobs/3948345
triggered by
https://github.com/eic/epic/actions/runs/11263980547/job/31323141281
just uses epic-main in the container, not the
update-sensor-geometry-for-RP it should be using.

This PR removes the direct trigger of the geoviewer.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: )
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>