Add configuration file for eHe3 (#807)
### Briefly, what does this PR introduce?
I would like to run simulation for e He3 scattering, but It seems like
the field was added for this configuration, but the .yml file was not.
So I added the configuration file for eHe3 beams.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

I hope I am submitting the pull request right this time.. Sorry for the
trouble!

Co-authored-by: Wan Lin <wlin3@eic0111.rcf.bnl.gov>