drop GeometryHelper, a copy of GeometryHelpers (#810)
### Briefly, what does this PR introduce?

Drops GeometryHelper which contains an unused copy of a subset of
GeometryHelpers. The only used function deals with DD4hep placement for
B0. GeometryHelpers is not a utility for DD4hep, but for geometry
calculations, so the code is moved to B0 unit.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
No

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>