add scaled field files for the early science magnet settings (#832)
### Briefly, what does this PR introduce?

This PR adds the necessary XML files to steer the beam correctly for 6
of the 7 early science settings. I have not added the 10x100 GeV Au
setting for right now since we are not even producing files for it.

[remedied with new commit] One thing that has NOT been added in the
configuration YAML files for craterlake. This is because those YAML
files don't have a field for the ion species.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ x] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No, it just adds the magnet settings to steer the ion beams for early
science.

### Does this PR change default behavior?

No, just adds additional functionality for ions.

---------

Co-authored-by: Alexander Jentsch <ajentsch@bnl.gov>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>