Bug fix on Barrel TOF sensors having incorrect segmentation.  (#834)
### Briefly, what does this PR introduce?

We realize that the cell positions are not in the correct places on some
BTOF sensors. Specifically, the cell center for the half-sensors
immediately to the left and right of the central gap for the engagement
ring are all incorrect. This issue is caused by the segmentation class
(CartesianGridXY) applying the wrong offset on the half-sensors.
Unfortunately, I do not know of a way to apply a different offset value
just for the half-sensor while keeping the original value for the normal
sensor.

This pull request resolves that issue by making all sensors half-sized.
A full sensor is simply two half-sensors placed side-by-side.

Please read our discussion on the ePIC AC-LGAD TOF weekly meeting for
details:
https://indico.bnl.gov/event/27008/contributions/103896/attachments/60237/103439/BTOF_geometry_bug_EIC_EPIC.pdf

### What kind of change does this PR introduce?
- [x ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [X ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

BTOF now uses CartesianStripX for segmentation rather than
CartesianGridXY. However, the field names for cellID remains unchanged
since I renamed the identifies for CartesianStripX to match the behavior
of CartesianGridXY.

### Does this PR change default behavior?

No.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>