Gapless beam pipe vacuum (#831)
### Briefly, what does this PR introduce?
Changes the beampipe chain so it no longer adds a small air gap when
there is a bend in the pipe and instead joins them together properly
with a torus segment.

Exaggerated geometry pictures showing the difference between current and
new implementation.

![Pipe_Chain_Old](https://github.com/user-attachments/assets/ac13485e-20b6-43a7-899b-ff8d44ed127b)

![Pipe_Chain_New](https://github.com/user-attachments/assets/35b04be2-0daa-4e1a-8804-66c07c2a37da)

The previous implementation was sufficient to remove the overlaps when
the bend has a very small angle and small radius, but a small air gap in
the vacuum isn't ideal even then.

Motivation for introducing this is that to update the backwards beamline
magnets to match the current lattice a bend needs to be added. Currently
I believe these bends in the beampipe chain are only present in material
we don't care about tracking particles through.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
Should remove small amounts of air from bending beamline chains with
bends.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>