add 10x130 interpolated beamline to enable normal usage like other en… (#839)
…ergies

### Briefly, what does this PR introduce?

This PR introduces a 10x130 lattice which has been calculated via
interpolation between the three "core" beam energy settings, while still
keeping the B0 constant. This will be paired with a similar PR for the
RP matrices in EICrecon.

This updated magnet setting file will allow for "normal" usage of the
file, where the B0 and RP can be used simultaneously.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ x] Tests for the changes have been added
- [x] Documentation has been added / updated
- [x ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No, improves performance.

### Does this PR change default behavior?

Yes, it changes the default 10x130 magnet settings to something which
can be properly used.