Geometry update for barrel TOF (#848)
### Briefly, what does this PR introduce?


### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

Geometry of BTOF stave and sensor design are updated to match the newest
proposal available here:
https://indico.bnl.gov/event/27154/contributions/104363/attachments/60759/104362/2025%2003%2024%20Mechanics%20Workshop%20TOF%20Presentation.pdf

I received a CAD drawing from the author of that presentation. I matched
epic geometry to the CAD drawing. This is how a stave looks like in CAD,


![image](https://github.com/user-attachments/assets/3e282818-dba6-4e35-a092-2b9828de2bc8)

Most prominently, the cooling tube design is modified and the stave's
length is reduced.

Here's the rendering of geoDisplay to show the updated BTOF stave. The
foam material for the stave on the left half is removed to show the
cooling pipes inside.

<img width="1042" alt="image"
src="https://github.com/user-attachments/assets/5ac275eb-1e21-4db7-a394-df83c980f4bd"
/>

### Does this PR change default behavior?

BTOF phase space coverage is reduced.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>