Centralize IP6 Beampipe constant definitions (#867)
### Briefly, what does this PR introduce?
Fixes an airgap in the Far-Backward beamline by centralizing the
definitions for the beampipe ends.

TaggerTrackerTracks have gone missing from the EICRecon output since the
IP6 beampipe update. A visual inspection shows an air gap between the
central beampipe and the current far-backward electron beamline (and an
overlap with the incoming hadron extension). The air is _hopefully_
responsible for breaking the electron track reconstruction...


![image](https://github.com/user-attachments/assets/ab40317b-7c43-4964-b7fb-ac77065f8e61)

Centralizing the variables used to define the positions of the ends
should fix this and stop something like this happening in the future.
The small gap in the incoming hadron beamline is much less important.


![image](https://github.com/user-attachments/assets/20dd30b3-2244-478a-8845-28f345747c55)

A more detailed revision of the far-backward beamline is in progress but
this is required sooner.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

### Does this PR change default behavior?
Hopefully recovers a full vacuum