Reduced grid size of ETOF for realistic position resolution. (#883)
### Briefly, what does this PR introduce?

The grid size for ETOF is incorrect. It should be 0.5mm on both x- and
y-direction. However, this pull request shrink the grid size all the way
to 0.1 mm because of the following:

A real LGAD sensor shares charge between neighbors. We are still working
on it and as such, charge sharing is not yet available for ETOF.
However, the position resolution of ETOF sensor is verified empirically
to be 0.03 mm when charge sharing is considered.

TOF group needs simulation results with the correct resolution.
Obviously the best way is the enable charge sharing and clustering in
EICrecon to recreate the observed 0.03 mm resolution, but the pull
request on LGAD clustering is still being reviewed and it doesn't look
like it will be approved soon.

As a result, our group decided that the easiest way to reduce position
resolution is to shrink grid size smaller than design specification. A
grid size of 0.1 mm gives us 0.1/sqrt(12) ~= 0.03 mm position.

We will revert back to realistic grid size once our charge sharing and
clustering classes are in place.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [x] Other: Improve ETOF resolution to something realistic.

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No

### Does this PR change default behavior?

No

---------

Co-authored-by: ShujieL <ShujieL@users.noreply.github.com>