fix: BarrelExtraSpace_thickness zeroed; DIRC support rails modified to avoid overlap (#894)
### Briefly, what does this PR introduce?
This PR removes 1.5 cm of extra space between the DIRC and the BIC. This
places the rmin of the BIC at 82.5 cm instead of 84 cm. This is closer
to how it should be, since in CAD it should be 82 cm. Despite the
remaining discrepancy, this should be merged.

This PR also hard-codes the carbon fiber inner detector support cylinder
radius in the definitions, so it can be used.

This PR also slightly modifies how the DIRC/MPGD support rails are
encoded. They now touch the inner detector support cylinder and the BIC
rmin.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No. This extra space only affects the EcalBarrel_rmin, and does not
propagate to other subsystems.

### Does this PR change default behavior?
Yes, BIC smaller.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>