feat: initial EcalBarrelImagingTracker support (#710)
### Briefly, what does this PR introduce?
This PR adds tracking support to the first AstroPix layer of the Barrel
Imaging Calorimeter (well, technically all of the layers).

With this geometry, and with https://github.com/eic/epic/pull/710, we
now see the hits from the first layer of the BIC be part of
reconstructed tracks (BIC first layer is avg at 84 cm rmin + 0.5 cm
inner support plate + 0.75 cm to middle of tray):

![image](https://github.com/user-attachments/assets/f206d580-1ec1-4c8e-9cf4-179a539aa7fd)

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Sylvester Joosten <sylvester.joosten@gmail.com>